Export onnx for model only (#361)
* Support export onnx for model only

* Fix

* Fix
