+++ projects/easydeploy/model/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .backendwrapper import BackendWrapper, EngineBuilder
+from .backendwrapper import BackendWrapper
 from .model import DeployModel
 
-__all__ = ['DeployModel', 'BackendWrapper', 'EngineBuilder']
+__all__ = ['DeployModel', 'BackendWrapper']
